Revert "Develop security"

+++ src/main/java/com/example/workshopmanagementapp/model/TaskService.java
@@ -34,10 +34,13 @@ public class TaskService {
 
     }
 
-    public void deleteById(Integer id) throws TaskNotFoundException {
+    public void deleteById(Integer id) throws TaskNotFoundException, MechanicNotEmptyException {
         Long count = taskRepository.countById(id);
         if (count==null || count.equals(0)){
             throw new TaskNotFoundException("Brak zadania o podanym ID");
+        }
+        if (!taskRepository.getById(id).getMechanic().equals(null)) {
+            throw new MechanicNotEmptyException("Zadanie ma przypisanego mechanika.");
         }
         taskRepository.deleteById(id);
